feat(core): revert back sort split into core sort
[COM-1202]
